copy the version file only if exists

+++ Scripts/build_kernel.sh
@@ -85,7 +85,9 @@ if [ "${BBG_SUPPORT}" = "true" ]; then
 fi
 
 cat "${EXTRA_CFG}" >> out/.config
-cp version out/version
+if [ -f version ]; then
+  cp version out/version
+fi
 make ${MAKE_ARGS} olddefconfig
 [ -f scripts/setlocalversion ] && sed -i 's/-dirty//g' scripts/setlocalversion || true
